Add new CheckList and make Changes.
