+++ build.gradle
@@ -62,6 +62,7 @@ ext.deps = [
         mockito              : '2.17.0',
         multidex             : '1.0.3',
         newrelic             : '5.14.0',
+        okhttp3              : '3.11.0',
         picassoTransforms    : '2.2.1',
         playServicesAnalytics: '16.0.1',
         powermock            : '2.0.0-beta.9',
