+++ library/base/src/main/java/org/cru/godtools/base/Settings.java
@@ -32,6 +32,7 @@ public final class Settings {
 
     @Nullable
     private static Settings sInstance;
+
     @NonNull
     public static Settings getInstance(@NonNull final Context context) {
         synchronized (Settings.class) {
