+++ ui/articles-aem-renderer/src/main/java/org/cru/godtools/articles/aem/view_model/ArticleViewModel.java
@@ -5,7 +5,6 @@ import android.arch.lifecycle.AndroidViewModel;
 import android.arch.lifecycle.LiveData;
 import android.support.annotation.NonNull;
 
-
 import org.cru.godtools.articles.aem.db.ArticleRepository;
 import org.cru.godtools.articles.aem.db.AttachmentRepository;
 import org.cru.godtools.articles.aem.db.ManifestAssociationRepository;
@@ -15,7 +14,6 @@ import org.cru.godtools.articles.aem.model.Attachment;
 import java.util.List;
 
 public class ArticleViewModel extends AndroidViewModel {
-
     private final ArticleRepository mArticleRepository;
     private final AttachmentRepository mAttchReposistory;
     private final ManifestAssociationRepository mManifestRepository;
@@ -32,7 +30,6 @@ public class ArticleViewModel extends AndroidViewModel {
         mArticleRepository.insertArticle(article);
     }
 
-
     public void deleteArticles(Article... articles) {
         mArticleRepository.deleteArticles(articles);
     }
@@ -40,7 +37,6 @@ public class ArticleViewModel extends AndroidViewModel {
     public LiveData<List<Article>> getArticles() {
         return mArticleRepository.getAllArticles();
     }
-
 
     public LiveData<List<Article>> getArticlesByManifest(String manifestID) {
         return mManifestRepository.getArticlesByManifestID(manifestID);
